Post school version update
This project is transitioning away from the school roots into it's own thing, so a lot of the older and less up to date documentation is being removed and/or marked as "old", until I can remove them.

Also tweaks the backend "getByName" functions to utilize query parameters for filtering, for use with the discord bot.